github-actions/requirements.txt
# Zscaler AI Guard — GitHub Actions policy validation
zscaler-sdk-python>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
